pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "backtester"
version = "0.1.0"
description = "Modular backtesting framework scaffold"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Julián", email = "juliandavid50327@gmail.com"}]
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.5", "mypy>=1.7"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
testpaths = ["tests"]
